backend/requirements.txt
Django>=4.2
djangorestframework
pandas
scikit-learn
joblib
